move competition model to shared folder
