+++ src/shared/models/competition.js
